move and scaladoc RequestFlowExtension
This commit
- adds ScalaDoc headers to the `RequestFlowExtension` types
- moves the two implementations into a separate package
    - this ensures that the API is "open enough" for external implementation
- drops the name `authenticationBackend` from the code; while authentication is
  the _raison d'être_ for the request flow extension, it's not the only
  conceivable use.
