+++ src/main/resources/reference.conf
@@ -37,7 +37,7 @@ druid = {
             max-connection-lifetime = 15 min
         }
 
-        authentication-backend = "ing.wbaa.druid.client.NoRequestFlowExtension"
+        authentication-backend = "ing.wbaa.druid.client.NoopRequestFlowExtension"
 
         authentication-config = { }
     }
