+++ src/main/scala/ing/wbaa/druid/auth/krb5/KDC.scala
@@ -1,0 +1,7 @@
+package ing.wbaa.druid.auth.krb5
+
+import scala.concurrent.Future
+
+object KDC {
+  def getTicket(): Future[String] = Future.successful("fake-ticket")
+}
