+++ src/main/scala/ing/wbaa/druid/client/RequestFlowExtension.scala
@@ -1,29 +1,65 @@
 package ing.wbaa.druid.client
 
-import scala.concurrent._
 import akka.http.scaladsl.model._
-import akka.http.scaladsl.model.headers._
-import com.typesafe.config.{ Config, ConfigFactory, ConfigValueFactory }
+import com.typesafe.config.{ Config, ConfigFactory }
 
+import scala.concurrent._
+
+/**
+  * Customization hook for altering the request flow in `DruidAdvancedHttpClient`. This is primarily
+  * intended to handle authentication requirements.
+  */
 trait RequestFlowExtension {
+
+  /**
+    * Modify the outgoing request before it is transmitted.
+    * @param request the outgoing request.
+    * @return the request that's to be sent to Druid.
+    */
   def alterRequest(request: HttpRequest): HttpRequest
+
+  /**
+    * Modify the incoming response before it's decoded. This method can be implemented to intercept
+    * an incoming authentication challenge, reach out to an authentication system such as a token
+    * issuer, and then re-transmit the original request to Druid with appropriate authentication
+    * headers.
+    *
+    * Implementors should be aware that any request sent through the provided `druidEndpoint` will
+    * itself pass through the request flow extension. Be careful not to infinitely recurse.
+    *
+    * @param request the original, outgoing request.
+    * @param response the response that came back from the server.
+    * @param druidEndpoint invoke to transmit a request to Druid.
+    * @param ec asynchronous context.
+    * @return the response future containing the answer to the original Druid query.
+    */
   def alterResponse(request: HttpRequest,
                     response: Future[HttpResponse],
-                    requestExecutor: HttpRequest => Future[HttpResponse])(
+                    druidEndpoint: HttpRequest => Future[HttpResponse])(
       implicit ec: ExecutionContext
   ): Future[HttpResponse]
 
+  /**
+    * Returns the active configuration of this extension.
+    */
   def exportConfig: Config
 }
 
+/**
+  * Marker trait for objects that produce a `RequestFlowExtension`.
+  */
 trait RequestFlowExtensionBuilder {
-
   def apply(config: Config): RequestFlowExtension
-
 }
 
-class NoRequestFlowExtension extends RequestFlowExtension {
+/**
+  * Forwards the request and response unmodified. Can be used as a null object, but is also suitable
+  * for use as a base class.
+  */
+class NoopRequestFlowExtension extends RequestFlowExtension {
+
   def alterRequest(request: HttpRequest): HttpRequest = request
+
   def alterResponse(
       request: HttpRequest,
       response: Future[HttpResponse],
@@ -34,85 +70,7 @@ class NoRequestFlowExtension extends RequestFlowExtension {
   override def exportConfig: Config = ConfigFactory.empty()
 }
 
-object NoRequestFlowExtension extends RequestFlowExtensionBuilder {
-  private lazy val instance                                = new NoRequestFlowExtension
+object NoopRequestFlowExtension extends RequestFlowExtensionBuilder {
+  private lazy val instance                                = new NoopRequestFlowExtension
   override def apply(config: Config): RequestFlowExtension = instance
-}
-
-class BasicAuthenticationExtension(username: String, password: String)
-    extends RequestFlowExtension {
-  def alterRequest(request: HttpRequest): HttpRequest =
-    request.withHeaders(Authorization(BasicHttpCredentials(username, password)))
-
-  def alterResponse(
-      request: HttpRequest,
-      response: Future[HttpResponse],
-      requestExecutor: HttpRequest => Future[HttpResponse]
-  )(implicit ec: ExecutionContext): Future[HttpResponse] =
-    response
-
-  override def exportConfig: Config =
-    ConfigFactory
-      .empty()
-      .withValue("username", ConfigValueFactory.fromAnyRef(username))
-      .withValue("password", ConfigValueFactory.fromAnyRef(password))
-}
-
-object BasicAuthenticationExtension extends RequestFlowExtensionBuilder {
-
-  override def apply(config: Config): RequestFlowExtension = {
-
-    val username =
-      Option(config.getString("username")).getOrElse {
-        throw new IllegalStateException(
-          "BasicAuthenticationExtension requires 'username' configuration parameter to be specified"
-        )
-      }
-
-    val password =
-      Option(config.getString("password")).getOrElse {
-        throw new IllegalStateException(
-          "BasicAuthenticationExtension requires 'password' configuration parameter to be specified"
-        )
-      }
-
-    new BasicAuthenticationExtension(username, password)
-  }
-}
-
-object KDC {
-  def getTicket(): Future[String] = Future.successful("fake-ticket")
-}
-
-class KerberosAuthenticationExtension extends RequestFlowExtension {
-  private var cachedTicket: Option[String] = None
-
-  def alterRequest(request: HttpRequest): HttpRequest =
-    cachedTicket
-      .map { ticket =>
-        request.withHeaders(Authorization(OAuth2BearerToken(ticket)))
-      }
-      .getOrElse(request)
-  def alterResponse(
-      request: HttpRequest,
-      response: Future[HttpResponse],
-      requestExecutor: HttpRequest => Future[HttpResponse]
-  )(implicit ec: ExecutionContext): Future[HttpResponse] =
-    response.map(_.status).flatMap {
-      case StatusCodes.Unauthorized =>
-        KDC.getTicket().flatMap { ticket =>
-          cachedTicket = Some(ticket)
-          requestExecutor(request)
-        }
-      case _ => response
-
-    }
-
-  override def exportConfig: Config = ConfigFactory.empty()
-}
-
-object KerberosAuthenticationExtension extends RequestFlowExtensionBuilder {
-
-  override def apply(config: Config): RequestFlowExtension = new KerberosAuthenticationExtension
-
 }
